\name{GLMaSPU-package}
\alias{GLMaSPU-package}
\alias{GLMaSPU}
\docType{package}
\title{
An Adaptive Test on High Dimensional Parameters in Generalized Linear Models
}
\description{
Provide various tests on high-dimensional parameters in generalized linear models.
}
\details{
Several tests for high dimensional generalized linear models have been proposed recently. In this package, we implemented a new test called aSPU for high dimensional generalized linear models, which is often more powerful than the existing methods in a wide scenarios. We also implemented permutation based version of several existing methods for research purpose. We recommend users use the aSPU test for their real testing problem.
}
\author{
Chong Wu, Wei Pan
Maintainer: Chong Wu <wuxx0845@umn.edu>
}

\references{
Chong Wu, Gongjun Xu and Wei Pan, "An Adaptive test on high dimensional parameters in generalized linear models" (Submitted)
}

\keyword{GLMaSPU}

