\name{GLMaSPU-package}
\alias{GLMaSPU-package}
\alias{GLMaSPU}
\docType{package}
\title{
\packageTitle{GLMaSPU}
}
\description{
\packageDescription{GLMaSPU}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{GLMaSPU}
}
\author{
Chong Wu, Wei Pan
Maintainer: Chong Wu <wuxx0845@umn.edu>
}

\references{
Chong Wu, Gongjun Xu and Wei Pan, "An Adaptive test on high dimensional parameters in generalized linear models" (Submitted)
}

\keyword{ GLMaSPU}

